refactor(frontend): split admin query/form logic into features
