+++ apps/frontend/features/admin/templates/types.ts
@@ -1,0 +1,25 @@
+export type TemplateAcceptType = 'file' | 'url';
+
+export type Template = {
+  id: string;
+  name: string;
+  description: string | null;
+  acceptType: TemplateAcceptType;
+  allowedExtensions: string[] | null;
+  urlPattern: string | null;
+  maxFileSizeMb: number;
+  isRequired: boolean;
+  sortOrder: number;
+  createdAt: unknown;
+};
+
+export type TemplateFormValues = {
+  name: string;
+  description: string;
+  acceptType: TemplateAcceptType;
+  allowedExtensions: string;
+  urlPattern: string;
+  maxFileSizeMb: number;
+  isRequired: boolean;
+  sortOrder: number;
+};
